chore: update ui (#776)
* chore: update ui

* chore: refactor proposal
