--- src/commands/dao-voting/proposal/slash.ts
@@ -1,1 +1,0 @@
-// TODO: slash version
