--- src/commands/dao-voting/voting/slash.ts
@@ -1,1 +1,0 @@
-// TODO: slash version
